+++ src/tools/document_ocr/__init__.py
@@ -1,0 +1,1 @@
+"""Document OCR tool package."""